active proposal template
